Add component and logic to give components elasticsearch access and apply it to the logcollector
This commit adds the logic that will create an es user for a component, create a secret with that es user and password in it for the component to use, and copies over the required elasticsearch secrets / config maps to give access the the component. This commit also applies this logic to the log collector controller.