Create Elasticsearch and Kibana hashes for LogStorage and use them in LogCollector
This commit does the following:
- creates hashes for Elasticsearch and Kibana (based off of their created at timestamps).
- Stores those hashes as annotations in the LogStorage object
- creates an annotation in the fluentd daemonset pod template equal to the Elasticsearch hash in LogStorage

The Elasticsearch and Kibana hashes used so that components dependent on Elasticsearch can have a way to roll over their resources if Elasticsearch or Kibana change in a way that a restart would be required. Right now the hash is only created from the created at timestamp of the Elasticsearch and Kibana resource as a complete rollover of these items is all we want to monitor right now.

We'll likely want to want to update the other components that rely on Elasticsearch or Kibana in a similar way in following PR's