RBAC changes for reports
tigera-ui-user was incorrectly given permissions to download global
reports. Removed such a permission. To be noted this is a change that
needs to be release noted as on upgrade admins part of that clusterrole
won't be able to download global reports anymore.

tigera-network-admin was incorrectly not given permissions to download
global reports. Added such a permission.