(SAAS-642) Generate certificate for compliance server and use it in voltron
Previously the compliance server was generating it's own certificate (within the code) if one wasn't provided. This PR generates that certificate, stores in a secret and mounts it to the compliance server.

Furthermore, we mount that secret to voltron so we can verify the certificate when voltron proxies requests to compliance.

Other changes in this PR:
- Remove configmap for guardian (we aren't allowing the proxy targets to be configurable and now it just doesn't seem needed for the other variables)