Merge Elasticsearch and Kibana current and new states properly
The ECK operator modifies the Elasticsearch and Kibana resources that we create, adding things like annotation and finalizers and in the Kibana CR, ECK also modifies the Spec. The tigera operator detects all changes to these CRs, and reconciles, attempting to write out the Elasticsearch and Kibana CR.

One caveat to the operators reconcilation of these resources is that it only writes out if the Spec of the CRs have changed. Since ECK modifies Kibanas spec, we attempt to change Kibana back, and this causes an update loop between the tigera operator and the eck operator.

One other problem that occurs is if we do actually want to update Elasticsearch or Kibana, we would write out the CRs without the annotations or finalizers that ECK added, so this triggers an immediate update from ECK to re add those annotations and finalizers. The operator would ignore updating Elasticsearch again, but this still isn't idea behaviour.

This commit copies over the Elasticsearch field in the kibana Spec, and the annotations / finalizers in the Elasticsearhc and Kibana spec from the current resource to the new resource we're updating kuberenetes with.