Don't overwrite resources in the status manager
If a controller calls CreateOrUpdate multiple times then the deployments, daemonsets, cronjobs, and statefulsets are overwritten which each call. This means that when syncState runs there may or may not be one of the resources required for the status manager to know if it should update the status (the status manager requires that it has one of the four previously mentioned resources to update the status)

This is particularily a problem for the logstorage resource, as it calls CreateOrUpdate multiple times, and not always with one of the four previously mentioned resources.

This commit modifies the Set<Resource> methods so that it adds stored resources to the status manager if those resources aren't already in there, instead of replacing the resources.