KUBERNETES_SERVICE_HOST/PORT override (#700)
* forward KUBERNETES_SERVICE_HOST/PORT to node ds

If there is no kube-proxy (or it is to be removed), e.g. when we enable
BPF, we need to override the KUBERNETES_SERVICE_HOST/PORT for operator
so that it can access k8s. The same setting needs to be forwarded to
other components like felix and cni.

* Add Typha

* render.K8sServiceEndpoint comments

* read KUBERNETES_SERVICE_HOST/PORT just once, at startup

(cherry picked from commit 8f0732fa1be6e86c7d5360dea425f56ac69478f8)