Degrade Calico status on typha autoscale issues
This commit ties the Calico tigera status to the typha autoscalar, degrading the status if there's an issue.

An error check that has been added is to make sure that there are enough linux nodes available to schedule the required typha pods on, and now if there's not, an error is logged and the Calico tigera status is set to degraded

If the auto scalar degraded the status, the core controller will see it if it reconciles and trigger the autoscalar run immediately. If this run fails the controller will set the degraded status again and try to re reconcile later. Note the the ticker in typha is reset if the run is triggered by the controller.