Add SupportedOSTypes function to Components
This forces every component to state what operating systems are supported for the objects being rendered. If a single OS is returned, then a node selector is used (to avoid any unnecessary disruption of the pods since node selectors were previously used), but if there are multiple OS types, then node affinity is used. The label for the selector / affinity is the "kuberentes.io/os" label.