Handle no resources to monitor in the status manager
This commit adds the ReadyToMonitor function to the StatusManager, and doesn't return false for IsAvailable if "progressing" and "failing" or nil. The purpose of ReadyToMonitor is to tell the status manager that it's OK to start reporting the available state of the resources it's been told about. If it hasn't been told about any resources, then the state is available (so long as there's no explicit degraded reason).