Add controller dependency graph
Add a dependency graph between different controllers in operator. This
is by analyzing the `Reconcile()` function in each controller. It is
part of the work from Operator Managed Prometheus.

This graph is generated using the python binding for graphviz. It is
still a manual process to update and regenerate the graph using the
python script `generate-controller-dependency-graph.py` at the moment.