Adding packetcapture api container for tigera-manager
PacketCapture Download API exposes the API to retrieve pcap files generated by a PacketCapture. This enables personas
like DevOps or Platform Engineers to retrieve files backed by a namespace RBAC.

PacketCapture API is a service that is deployed as container part of tigera-manager deployment. By default, it is
installed by Tigera operator. This service will only be deployed only for Standalone and Management clusters, and requests
will be routed to the appropriate managed clusters.

Voltron container has route defined for any requests that starts with `/packet-capture/` prefix. It will strip away that
prefix and forward it to the container where this service resides.