Add a packet capture container to tigera-guardian
Add a new container for packet capture on guardian side. In a
multicluster setup, the management cluster will proxy a request via the
tunnel to Guardian. Guardian will forward the request to the packet
capture container. This API will retrieve the files generated by a
packet capture via a pod/exec command.