Correct Elasticsearch metrics ServiceMonitor name
This typo in Elasticsearch metrics ServiceMonitor name is introduced in
v3.8 and will be corrected in v3.9. The old ServiceMonitor object will
be dangled in the cluster so a `toDelete` object is added. This object
should be deleted after we drop support for v3.8.

This changeset also fix some unit test issues after merging from the
latest operator master.