Move misc Prometheus resources to Tigera Operator
The key change for this PR is to move Alertmanager configuration secret
to be managed by the Operator. The secret will be created first in the
Operator namespace (for new installation) or copied from the Prometheus
namespace (for existing installation / upgrade). Changes in the Operator
namespace will be propogated to the Prometheus namespace. Changes in the
Prometheus namespace will be reverted.

The following Prometheus resources are moved to the Operator after this
PR:
* 99-alertmanager-secret.yaml
* 99-alertmanager-service.yaml
* 04-clusterrole-prometheus.yaml
* 04-clusterrolebinding-prometheus.yaml
* 04-serviceaccount-prometheus.yaml