Get internal elasticsearch secret from proper namespace
The internal es secret is no longer in the operator namespace since other components don't use it (they use the gateway secret). The only secret that needs it now is the gateway.

This also fixes a bug where we're not checking if the kibana or elasticsearch internal secrets are nil, meaning they don't exist (because GetSecret returns a nil secret with no error if the secret doesn't exist).