Remove "duplicate" image-all target (use images-all instead)
This target doesn't seem to build the arm64 image, it just hangs. The might be some oddness with the sub-image-% target being redeclared that's causing this, but all I know is that images-all works and it's used in the CI but image-all doesn't and it's only used for the release building.