Add the tenant ID as the es midfix for compliance server
We've made a change to compliance so that you can add a midfix and a
suffix. For the compliance server the cluster id is taken from the
request (X_CLUSTER_ID) to download compliance reports and the es suffix
is overwritten (which has the tenant ID in it). We will set the midfix
to the tenant ID and concatenat that with the X_CLUSTER_ID suffix to
download and list reports for a managed cluster.