adding changes to install image assurance - CNX-15419
This PR adds a new CRD and controller for imageassurance. Existence of the
tigera-secure ImageAssurance CR enables the feature. It installs API and scanner
components needed for image assurance.

Both API and Scanner communicate with backend PostgreSQL database
hosted on a devops cluster, changes for this communication are include in
this PR. In addition to database API communicates
with tigera manager via voltron proxy. (changes for this will be
addressed in a seperate PR)