changes to manager for communicating with image assurance api- CNX-15616
This PR includes changes for manager and voltron to communicate with
image assurance api. Manager UI proxies the image assurance requests
(using the the path /image-assurance/) to voltron (proxy). Voltron sends the
requests to image assurance service and sends back the response to frontend.

For this communication to happen we need to mount image assurance api tls
cert to voltron and also enable some env vars containing config for
image assurance for both manager and voltron. This is a follow up of imageassurance
installation PR #46 tracked by CNX-15419.