Add AssignIpv[46] in CalicoNetworkSpec
This patchs adds two additional parameters to the CalicoNetworkSpec
allowing to force the value of assign_ipv4 / assign_ipv6 in the CNI
configuration. This is useful when the ippools are not created at
installation time but later on during provisionning.

If omitted, the previous behaviour will be preserved (deriving the
values from the presence of v4/v6 pools in the default ippools)

Signed-off-by: Nathan Skrzypczak <nathan.skrzypczak@gmail.com>