Add support for component resource overrides (#2063)
* Add new fields to Installation to allow component resource customization:
- CalicoNodeDaemonSet
- CalicoKubeControllersDeployment
- TyphaDeployment
- CalicoWindowsUpgradeDaemonSet

* Add APIServerDeployment field to APIServerSpec

* Add k8s validation helpers

* Replace 'kubectl apply' for operator CRDs with 'kubectl create'

Now that we've added more fields to the Installation API, we are hitting
the error: The CustomResourceDefinition "installations.operator.tigera.io" is invalid: metadata.annotations: Too long: must have at most 262144 bytes

* Use same calico API server container name for OSS and Enterprise

This allows us to use a single API server container name for the
APIServerDeployment configuration for both OSS and Enterprise.
This makes upgrades easier from OSS to Enterprise

* Use the same apiserver CSR init container name for OSS and Enterprise

This allows us to use a single API server CSR init container name for the
APIServerDeployment configuration for both OSS and Enterprise.
This makes upgrades easier from OSS to Enterprise

* Merge provided nodeSelector with default nodeSelector

Instead of replacing default nodeSelector with provided nodeSelector,
"merge" the key/value pairs with the existing default nodeSelector.
Only keys of the provided nodeSelector key/value pairs that don't
already exist in the default nodeSelector are added.

This behaviour is inline with how ControlPlaneNodeSelector works today.