[IA-417] - Add new component runtime cleaner (#210)
* Add working runtime support with tests

This PR adds a new image assurance component with the required k8s components. The runtime cleaner is a deployment that resides in the management cluster and cleans up stale data that customers' have in the Image Assurance Database. It polls the customers' managed clusters every few minutes (configurable) and deletes any stale nodes and clusters in our database.