update IA api access token util, update test cases (#250)
Starting from v1.24 kubernetes will no longer auto generate API token (item 3) for service accounts, we have to explicitly create a token for each service account that is not mounted on pods. Kube-controller does it now with tigera/calico-private#5749

This PR makes changes to operator rendered components that use the token. Instead of accessing SA and then secrets, we ensure SA exists and then directly access the secret with the well known name. It also adds watches for the new secrets and updates the test cases.