Merging changes for adding CR adaptor for image assurance (#253)
This PR includes the following changes.

With this we are merging all the changes we previously reviewed/merged in to image-assurance-dev into master (#210 #230 #234 #237 #231 #245 #248 #250 #252). We created a separate branch (image-assurance-dev) to be able to merge the PRs related to CR adaptor into the repo without merging into master (as we were cutting a release at the time and these changes aren't scheduled to go into the release that was being cut).

This PR adds a new image assurance component with the required k8s components. The runtime cleaner is a deployment that resides in the management cluster and cleans up stale data that customers' have in the Image Assurance Database. It polls the customers' managed clusters every few minutes (configurable) and deletes any stale nodes and clusters in our database.

This PR also includes changes for RBAC changes. Starting from v1.24 kubernetes will no longer auto generate API token (item 3) for service accounts, we have to explicitly create a token for each service account that is not mounted on pods. Kube-controller does it now with tigera/calico-private#5749. This PR makes changes to operator rendered components that use the token. Instead of accessing SA and then secrets, we ensure SA exists and then directly access the secret with the well known name. It also adds watches for the new secrets and updates the test cases.

This PR also includes changes to sync imageassurance config map periodically and sync it with the settings we have in Databse.

* Add cluster role for operator's access to the image assurance API

* Add permission to get pods from IA API when cleaning up at startup (#234)

* Remove Pod Watcher (CR Adaptor running in managed clusters replaces this)

* Sync cache IA API settings (from db) in image assurance config map (#231)

* [IA-417] - Add new component runtime cleaner (#210)

* Add working runtime support with tests

* use common function to access image assurance token (#248)

* update IA api access token util, update test cases (#250)

* update image-assurance-dev with master (#252)

* TSLA-3828 voltron configuration parameter CheckManagedClusterAuthorizationBeforeProxy is not enabled (#246)

* TSLA-3828 voltron configuration parameter CheckManagedClusterAuthorizationBeforeProxy is not enabled

* removing noisy log

* adding review changes for image-assurance-dev to master merge  (#254)

* adding review changes

* adding review changes

* run make gen-versions