Set DNS trusted servers using FelixConfig instead of FELIX_ env var
For Rancher and OpenShift the operator needs to set a non-default value for DNS trusted servers, and
it was previously doing that by setting a FELIX_ env var in the node daemonset.  But that makes it
impossible for any further tweaks to that using FelixConfiguration - because FelixConfiguration
settings are always overridden by FELIX_ env vars.

To allow later tweaks, change the operator to create or patch FelixConfiguration instead.

References:
https://tigera.atlassian.net/browse/RS-897
https://tigera.atlassian.net/browse/EV-3193