Update K8s pins to 0.27.9 and controller-runtime to 0.15.3 / fix incompatibilities
This commit updates the pins to 0.27.9 (and controller-runtime to 0.15.3, where most of the incompabitilities lie)  and fixes the incompatibilities
with the new versions, specifically:
- updating the manager creation, replacing the removed cache builder and mapper
  provider inputs with the replacements
- Implementing a new Controller layer to cache the manager cache to deal
  with the new source.Kind requirement where you must provide a cache to
  the source.Kind function (previously a struct) when creatin a watch
- Replace bits of deprecated functionality from the apiserver wait
  package.
- Automatically register types that have statuses with the fake client
  using utility function (types with status now must be registered)
- updated egress controller test to not wait for proper calico version,
  just expect the first reconcile to fail and the next one to succeed
  after setting the right version

For the second bullet, I've introduced our own Controller interface
which extends the controller.Controller interface so the individual
controller / controller utilities don't need to deal with cache management.
As outlined in some of the comments for that package, I intendd to use
this layer to house future adaptations for cache management (and
possibly even multi tenancy) to the controller runtime package.