Fix autodetection of current RKE2
Previously based on the kube-system/rke2 ConfigMap, but in current RKE2 the
above ConfigMap no longer exists.  We leave that code in place in case there
are variants where it is useful, but check also for the RKE2 DNS service -
which is especially relevant because one of the main uses of the RKE2
autodetection is to set DNS config.