Disable keep alive for the elasticsearch client
We seem to be leaking connections because we're not reusing the http
client for elasticsearch. The http client by default keeps connections
around (so they can be reused), even if you're not keeping the http
client around.

This commit disables keeping connections alive for the client. This is
preferable to trying to reuse the client because we're already trying to
use this as a single use client, and changing the code to reuse the
client isn't very beneficial to us as this is just for start up
functionality.