Move test utilities to test package fromt utils folder to move ginkgo v1 import
Currently we import ginkgo v1 in the utils.go folder for test utilities.
This means that anything importing the controllers package also
imports ginkgo v1, and if that importer also imports ginkgo v2 then a
panic will occur.

Moving the test utilities to the test folder ensures we don't import
ginkgo v1 when we're using the controller libraries.