[CORE-10407] Add support for OpenShift HCP (#3389)
* [CORE-10407] Add support for OpenShift HCP

Add support for OpenShift HCP (Hosted Control Planes). Main change is on
the AWS security group setup job, as the SG names are different in an
HCP hosted cluster as opposed to vanilla OpenShift (or a hosting cluster
which is essentially the same). Also added auto-detection for an HCP
hosted cluster (when ControlPlaneTopology == "External") and propagation
of this info to the SG setup job via an env var.

* make 'gen-versions'

* Use cached client to verify if on hosted OpenShift and address other review comments

* Add watch for OpenShift infrastructure config