Fix image component handling, honoring Installation spec
components.GetReference currently ignores the image prefix and path
specified in the "Installation" spec.

It makes a copy of the image component, appends the configured image
path and prefix but then it uses the original component image when
looping over the image set.

The fix is trivial, we just have to use the modified image string.