Reduce the number of restart for tigera-manager (#3503)
Tigera-manager is restarted by operator in a multi-tenant environment
due to the following:
- we are picking up kibana certificates instead of elastic and we are
setting the annotations to use kibana certificate. At the next
reconciliation, this changes. This is happening because external kibana
and external elastic have the value and thus the same hash. Certificate
manager only keeps one of them, as it stores as a hash for the
certificate. We do not need external kibana certificates and we can stop
rendering them
- voltron routes are configured by different components (cloud rbac,
image assurance) and can produce different volume mounts/volumes which
causes tigera-manager to create a new replica set.