Fix for PriorityClassName override handling
This is a follow-up to #3613.  During the work in #3626 to add some additional
UT for that, I discovered the problem that is fixed here.  `getField` can
return:
1. an invalid value - when the field being looked for doesn't exist
2. a nil pointer or slice value - when a field along the looked for path exists but is nil
3. a string value - when everything exists and ends up at a string field.

In case (2), `value.String()` will - misleadingly for our purposes - return a
non-empty string representation of the fact that the value is a nil pointer.
But what we really want in that case is just "".  So make sure we only use
`value.String()` when `value.Kind()` is `String`, and otherwise return "".