[release-v1.34] Auto pick #3389: Add support for OpenShift HCP #3640: Update AWS SG setup job for OpenShift (#3642)
* Add support for OpenShift HCP

Add support for OpenShift HCP (Hosted Control Planes). Main change is on
the AWS security group setup job, as the SG names are different in an
HCP hosted cluster as opposed to vanilla OpenShift (or a hosting cluster
which is essentially the same). Also added auto-detection for an HCP
hosted cluster (when ControlPlaneTopology == "External") and propagation
of this info to the SG setup job via an env var.

* make 'gen-versions'

* Use cached client to verify if on hosted OpenShift and address other review comments

* Add watch for OpenShift infrastructure config

* Update AWS SG setup job for OpenShift

* make fix

* Fix bad refactor on AWS SG setup (#3645)

* Fix bad refactor on AWS SG setup

Fix refactor made in https://github.com/tigera/operator/pull/3633 that
did not consider SG rules between SGs. This adds a string slice argument
to setupSG() so that multiple SG IDs can be passed and set up as src for
the rules.

* make fix