Create CRDs if they don't already exist, but never update existing ones
Ref. https://github.com/tigera/operator/pull/3638#discussion_r1884181047

The point here is to avoid any possible fighting between our operator and
another component that might also be trying to install CRDs for the Gateway
API, e.g. if a customer is using multiple Gateway API implementations at the
same time.

This change only stops us from fighting.  It's possible to imagine more
sophisticated logic, such as installing _our_ CRDs if they have versions which
are later (or in some way "better") than the existing ones.  But it's not
obvious that that would be what our customers want - e.g. they might _need_ an
older version - and even if it wasn't we could still be fighting with a less
intelligent other component.