fix unstable Expect().To() for large nested dex policies
gomega Expect was not able to compare correctly the dex policies
and would fail when running in "Run" mode but pass when running in
"Debug" mode.
It was also passing when we print the values before the Expect().To()
call.

To fix this, we convert the larger and nested struct to String
and compare the string values.