Allow kube-controllers to read its own configuration
Previously we had optimisations in libcalico-go that translated a list of one
item into a get.  But that wasn't sound and was removed in #8469.  So there can
be places where we were previously OK with a "get" permission but now need
"list" as well.

Specifically in the latest 3.21 hashrelease ("snowdrift"), we see:

    2025-01-15 16:51:10.404 [WARNING][13] kube-controllers/runconfig.go 219: unable to list KubeControllersConfiguration(default) error=connection is unauthorized: kubecontrollersconfigurations.crd.projectcalico.org "default" is forbidden: User "system:serviceaccount:calico-system:calico-kube-controllers" cannot list resource "kubecontrollersconfigurations" in API group "crd.projectcalico.org" at the cluster scope