Update libcalico-go version to the branch, rather than the last released tag
As a consequence of fixing libcalico-go being pinned to the latest tag, we do not update CRD changes between that tag and the branch, which is
problematic for the new feature where the operator deploys CRDs itself. ECK therefore, was still pinned to ECK 2.6.1, whereas we are currently
deploying v2.16.0, due to CVE fixes, etc.