Update Envoy gateway code from v1.1.2 to v1.2.6 (#3731)
* Update Envoy gateway code from v1.1.2 to v1.2.6

This is to address CVE-2025-24030
https://github.com/envoyproxy/gateway/security/advisories/GHSA-j777-63hf-hx76

As part of this:

- I added a handy 'mod-tidy' target in the Makefile.

- Various updates needed for version bumps in some of the k8s libraries that we
pull in.

* Update NewIndexerInformer to NewInformerWithOptions

* FV: Skip controller name validation

Moving from controller-runtime v0.18.5 to v0.19.1, by default there is [a new
global
variable](https://github.com/kubernetes-sigs/controller-runtime/blob/v0.19.1/pkg/controller/name.go#L27)
checking against reuse of controller names.  That scuppers our FV tests because
they all run in a single process execution.  But there's a SkipNameValidation
option that we can use to disable that checking.