[RS-2260] Add GatewayAPI field to configure CRD management
The default behaviour - which is used when the new field is not set, or is set
to "Reconcile" - is to keep the cluster's Gateway API CRDs aligned with those
that the Tigera operator would install on a cluster that does not yet have any
version of those CRDs.

Alternatively, if this field is set to "PreferExisting", the Tigera operator
will create the Gateway API CRDs if they do not already exist, but will not
overwrite any existing Gateway API CRDs.  This setting may be preferable if the
customer is using other implementations of the Gateway API concurrently with
the Gateway API support in Calico Enterprise.  It is then the customer's
responsibility to ensure that CRDs are installed that meet the needs of all the
Gateway API implementations in their cluster.