Add guardian to whisker (#3793)
This commit adds guardian to the whisker deployment, and is only launched if the ManagementClusterConnection CR exists.

This adds the ManagementClusterConnection CRD to open source.

Guardian might not stay in this deployment (and likely won't), but it's a good temporary place for it alongside goldmane (where the service isn't exposed).