Add rbac to delete secrets for external controller
When connecting a managed cluster to a management single tenant cluster,
es-kube-controllers are responsible for copying the voltron linseed
certificate via the tunnel. Before this operation, es kube controllers
will first reconcile users. Any failure in reconciling users will lead
to the certificate not being copied over. The first step in user
reconciliation is to delete secret for decommisioned users or
components, like curator. A failure because of missing RBAC will result
the certificate not copied over, namespaces for fluentD, commpliance and
intrusion controller not being created in the managed cluster.