Bugfix: Provision EnvoyProxy with the gateway class name (#4019)
There was a bug in the logic of cleaning up unwanted GatewayClasses and
EnvoyProxies: if a custom EnvoyProxy had a name different from that of the
GatewayClass using it, the renderer would accidentally delete the EnvoyProxy
again from the tigera-gateway namespace, almost immediately after provisioning
it, making the GatewayClass inoperable.

As part of this fix, we now create EnvoyProxies in the tigera-gateway namespace
with the same name as that of the GatewayClass.  In other words:

    custom EnvoyProxy exists with <user-name> in <user-namespace>

    GatewayClasses[*].Name = <custom-class-name>
    GatewayClasses[*].EnvoyProxyRef = {<user-name>, <user-namespace>}

    =>

    EnvoyProxy provisioned with <custom-class-name> in tigera-gateway
    GatewayClass provisioned with <custom-class-name> in tigera-gateway

whereas previously the penultimate line would have been

    EnvoyProxy provisioned with <user-name> in tigera-gateway

My previous thinking was "if the user has chosen a name, let's respect that."
My thinking now is:

1. That leads to bugs like the one here.

2. If we respect `<user-name>`, it's possible that `<user-name>` for the
EnvoyProxy of a GatewayClass with EnvoyProxyRef will clash with
`<custom-class-name>` for another GatewayClass without EnvoyProxyRef.  Let's
just use `<custom-class-name>` in all cases.