Gateway API: allow for OpenShift 4.19 CRD lockdown
OpenShift 4.19+ pre-installs some of the Gateway CRDs, but not all of them, and
has a webhook that prevents us from installing the ones that OpenShift is
missing.  To work with this we distinguish between "essential" and "optional"
CRDs.  The "essential" set must be a subset of those that OpenShift installs
and/or allows to be installed, and must also suffice for all of the
Gateway-related feature that we consider important as part of Calico; and this
controller will report an error and degraded status if any of those do not
already exist and cannot be installed.  The "optional" set is everything else
that we would ideally install, to provide more options to our users; but this
controller will only warn if any of those cannot be installed (and do not
already exist).

Fixes https://tigera.atlassian.net/browse/RS-2689
Fixes https://tigera.atlassian.net/browse/RS-2690