[release-v1.38] Auto pick #4137: feat(dex): Enable PKCE and silent token renewal for SPA (#4138)
* feat(dex): Enable PKCE and silent token renewal for SPA

This change updates the Dex client configuration to support Single Page Applications (SPAs).

The Dex client is configured as public (), which disables the use of a client secret and enables the more secure Proof Key for Code Exchange (PKCE) flow. This is the recommended standard for public
clients like browser-based applications.

Additionally, the 'X-Frame-Options' header is changed from 'DENY' to 'SAMEORIGIN'. This allows the SPA to use a hidden iframe to perform silent OIDC callbacks, enabling seamless token renewal without requiring a full-
page redirect.

Together, these changes improve security and user experience for the Calico Enterprise Manager when used as an SPA.

* feat(dex): Enable PKCE and silent token renewal for SPA

This change updates the Dex client configuration to support Single Page Applications (SPAs).

The Dex client is configured as public (`public: true`), which disables the use of a client secret and enables the more secure Proof Key for Code Exchange (PKCE) flow. This is the recommended standard for public
clients like browser-based applications.

Additionally, the 'X-Frame-Options' header is changed from 'DENY' to 'SAMEORIGIN'. This allows the SPA to use a hidden iframe to perform silent OIDC callbacks, enabling seamless token renewal without requiring a full-
page redirect.

Together, these changes improve security and user experience for the Calico Enterprise Manager when used as an SPA.

* feat(dex): Enable PKCE and silent token renewal for SPA

This change updates the Dex client configuration to support Single Page Applications (SPAs).

The Dex client is configured as public (`public: true`), which disables the use of a client secret and enables the more secure Proof Key for Code Exchange (PKCE) flow. This is the recommended standard for public
clients like browser-based applications.

Additionally, the 'X-Frame-Options' header is changed from 'DENY' to 'SAMEORIGIN'. This allows the SPA to use a hidden iframe to perform silent OIDC callbacks, enabling seamless token renewal without requiring a full-
page redirect.

Together, these changes improve security and user experience for the Calico Enterprise Manager when used as an SPA.