Fix OIDC PKCE flow + Google connector (#4159)
* fix: fix email unverified issue when obtaining dex token

In recent times, the Google flow no longer results in email_verified claim being true by default. We now propagate the insecureSkipEmailVerify setting also for the google connector.

* fix!: add missing offline_access claim to requestedScopes when not defined by user

Although documented in our api as a default value when requestedScopes is not provided, the code no longer defaulted this value since PR #1065. Likely, this has to do with the fact that only recently we started using this claim when we switched from the implicit flow to the code flow for our UI.

BREAKING CHANGE: fixed the defaulting behaviour for Authentication.Spec.OIDC.requestedScopes such that it now includes offline_access as documented in the API.