Ensure dns names ending with a "." aren't added to certificates
If the clusterDomain isn't set then we add domain names ending with a ".", which are invalid. Golangs crypto library just started enforcing valid dns names, and our fv tests are failing because we don't have a cluster name set and it results in entries like "calico-api.calico-system.svc." (notice the end dot!!) being set in the certificate.